+++ src/main/java/hu/multiteam/papyrum/block/PapyrumBlocks.java
@@ -5,6 +5,7 @@ import hu.multiteam.papyrum.item.PapyrumItems;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.RotatedPillarBlock;
 import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraft.world.level.material.MapColor;
 import net.minecraftforge.eventbus.api.IEventBus;
@@ -16,7 +17,9 @@ import java.util.function.Supplier;
 
 public class PapyrumBlocks {
     public static final DeferredRegister<Block> BLOCKS = DeferredRegister.create(ForgeRegistries.BLOCKS, Papyrum.MODID);
-    public static final RegistryObject<Block> PAPER_BLOCK = register("paper_block", () -> new Block(BlockBehaviour.Properties.of().mapColor(MapColor.TERRACOTTA_WHITE)), new Item.Properties());
+    public static final RegistryObject<RotatedPillarBlock> PAPER_BLOCK = register("paper_block", () -> new RotatedPillarBlock(BlockBehaviour.Properties.of().mapColor(MapColor.TERRACOTTA_WHITE)), new Item.Properties());
+    public static final RegistryObject<RotatedPillarBlock> CARDBOARD_BOX = register("cardboard_box", () -> new RotatedPillarBlock(BlockBehaviour.Properties.of().mapColor(MapColor.COLOR_BROWN)), new Item.Properties());
+    public static final RegistryObject<Block> CARDBOARD_BLOCK = register("cardboard_block", () -> new Block(BlockBehaviour.Properties.of().mapColor(MapColor.COLOR_BROWN)), new Item.Properties());
 
     public static void register(IEventBus bus){
         BLOCKS.register(bus);