+++ changelog.md
@@ -3,3 +3,4 @@
 - added cardboard item
 - added paper block
 - added recipe for the paper block
+- added cardboard box block